Files Section of ISS

Files to be installed on user's computer. Everything in app_dir plus file_list. For more information, visit [Files] section.

files_section(iss, app_name, app_dir, user_browser,
  file_list = character())

Arguments

iss
Character vector which cumulatively becomes an Inno Setup Script (ISS).

app_name
The name of the app. It will be displayed throughout the installer's window titles, wizard pages, and dialog boxes. See [Setup]:AppName for details. For continuous installations, app_name is used to check for an R package of the same name, and update it. The Continuous Installation vignette has more details.

app_dir
Development app's directory, defaults to getwd().

user_browser
Character for the default browser. Options include "chrome", "firefox", and "ie."

file_list
Character vector. Extra files to be installed with the app.

Value

Chainable character vector, which can be used as the text argument of writeLines to generate an ISS.